sync-service/requirements.txt
pymongo==4.6.0
elasticsearch==8.11.0
qdrant-client==1.7.0
sentence-transformers>=2.3.0
torch>=2.0.0
numpy==1.24.3
pyyaml==6.0.1
python-dateutil==2.8.2
schedule==1.2.0
tenacity==8.2.3
loguru==0.7.2
pydantic==2.5.0
tqdm==4.66.1

